pyproject.toml
[tool.poetry]
name = "routes4life-api"
version = "0.1.0"
description = ""
authors = ["Alexey Lavrenovich <lavrenovichae@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.2"
djangorestframework = "^3.13.1"
django-environ = "^0.8.1"
python-decouple = "^3.6"
djangorestframework-simplejwt = "^5.0.0"
djangorestframework-camel-case = "^1.3.0"
psycopg2-binary = "^2.9.3"
gunicorn = "^20.1.0"
pytest = "^7.0.1"
pytest-django = "^4.5.2"
Faker = "^13.2.0"
factory-boy = "^3.2.1"
pytest-factoryboy = "^2.1.0"
redis = "^4.2.2"
django-redis = "^5.2.0"
boto3 = "^1.23.4"
Pillow = "^9.1.1"
django-storages = "^1.12.3"
djangorestframework-gis = "^1.0"
geopy = "^2.2.0"
postgis = "^1.0.4"
mailchimp-marketing = "^3.0.75"
sendgrid = "^6.9.7"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
isort = "^5.10.1"
pylint = "^2.13.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
